+++ Editor/Interfaces/IPlantUmlResult.cs
@@ -1,4 +1,4 @@
-namespace asmdef2pu
+namespace asmdef2pu.Interfaces
 {
     internal interface IPlantUmlResult
     {
